Initial version of the program
Uses JGit for pure-Java Git functionality (not fully implemented yet,
though), and args4j for parsing command line arguments.

This first version implements logic for parsing an image file, rescaling
it to a 2-bit 54x7 matrix, computing the number of contributions needed
for each day, and simulating the generation thereof, visualizing the
process in a custom Swing panel.

Still to do:
  * Create the output Git repository to match the needed contributions.
  * Use ImageJ2 to provide a GUI for harvesting inputs, if --gui given.
  * Parse existing adjustment repository for data that would be
    overwritten by force pushing the new output repository, and factor
    that in to the computation.
  * Match the shades of green more closely to the GitHub web site.