Use more efficient commit counts
There is no reason to scale every pixel to the maximum value of each
desired bin. For example, for a scale factor of 15, we were scaling as
follows:

  0 -> 60 [bin 46-60]
  1 -> 45 [bin 31-45]
  2 -> 30 [bin 16-30]
  3 -> 15 [bin  1-15]

When we could be scaling:

  0 -> 46 [bin 46-60]
  1 -> 31 [bin 31-45]
  2 -> 16 [bin 16-30]
  3 ->  1 [bin  1-15]

The only exception is that we must have *one* pixel with a value of 60,
so that the bins still get defined as stated above.